requirements.txt
cassandra-driver
elasticsearch
PyYAML
futures # only for py2
